pyproject.toml
[project]
name = "polids"
version = "2.0.0"
description = "Analysis of political data and output of it through apps."
authors = [{ name = "AndreCNF", email = "andrecnf@gmail.com" }]
license = "MIT"
readme = "README.md"
requires-python = ">=3.13, <4.0"
dependencies = [
    "docling>=2.28.0",
    "google-genai>=1.7.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "langfuse>=2.60.2",
    "lingua-language-detector>=2.1.0",
    "llama-cloud-services>=0.6.8",
    "llama-index-core>=0.12.26",
    "llama-index-readers-file>=0.4.7",
    "loguru>=0.7.3",
    "marker-pdf>=0.2.17",
    "matplotlib>=3.10.1",
    "nbformat>=5.10.4",
    "nltk>=3.9.1",
    "ollama>=0.4.7",
    "openai>=1.68.2",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "plotly>=6.0.1",
    "pycountry>=24.6.1",
    "pydantic-extra-types>=2.10.4",
    "pydantic-settings>=2.8.1",
    "pymupdf>=1.25.4",
    "pymupdf4llm>=0.0.18",
    "python-dotenv>=1.1.0",
    "sentence-transformers>=4.0.1",
    "simplemma>=1.1.2",
    "wordcloud>=1.9.4",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
]
